Package refactoring. Introducing "internal" package.
