+++ flink-connector/src/main/java/io/delta/flink/source/DeltaSourceBuilder.java
@@ -8,13 +8,15 @@ import io.delta.flink.source.DeltaSourceBuilderSteps.MandatorySteps;
 import io.delta.flink.source.DeltaSourceBuilderSteps.TableColumnNamesStep;
 import io.delta.flink.source.DeltaSourceBuilderSteps.TableColumnTypesStep;
 import io.delta.flink.source.DeltaSourceBuilderSteps.TablePathStep;
-import io.delta.flink.source.enumerator.BoundedSplitEnumeratorProvider;
-import io.delta.flink.source.enumerator.ContinuousSplitEnumeratorProvider;
-import io.delta.flink.source.exceptions.DeltaSourceException;
-import io.delta.flink.source.exceptions.DeltaSourceExceptionUtils;
-import io.delta.flink.source.file.AddFileEnumerator;
-import io.delta.flink.source.file.DeltaFileEnumerator;
-import io.delta.flink.source.state.DeltaSourceSplit;
+import io.delta.flink.source.internal.DeltaPartitionFieldExtractor;
+import io.delta.flink.source.internal.DeltaSourceOptions;
+import io.delta.flink.source.internal.enumerator.BoundedSplitEnumeratorProvider;
+import io.delta.flink.source.internal.enumerator.ContinuousSplitEnumeratorProvider;
+import io.delta.flink.source.internal.exceptions.DeltaSourceException;
+import io.delta.flink.source.internal.exceptions.DeltaSourceExceptionUtils;
+import io.delta.flink.source.internal.file.AddFileEnumerator;
+import io.delta.flink.source.internal.file.DeltaFileEnumerator;
+import io.delta.flink.source.internal.state.DeltaSourceSplit;
 import org.apache.flink.configuration.ConfigOption;
 import org.apache.flink.connector.file.src.assigners.FileSplitAssigner;
 import org.apache.flink.connector.file.src.assigners.LocalityAwareSplitAssigner;
@@ -24,13 +26,13 @@ import org.apache.flink.table.data.RowData;
 import org.apache.flink.table.types.logical.LogicalType;
 import org.apache.flink.table.types.logical.RowType;
 import org.apache.hadoop.conf.Configuration;
-import static io.delta.flink.source.DeltaSourceOptions.IGNORE_DELETES;
-import static io.delta.flink.source.DeltaSourceOptions.PARQUET_BATCH_SIZE;
-import static io.delta.flink.source.DeltaSourceOptions.STARTING_TIMESTAMP;
-import static io.delta.flink.source.DeltaSourceOptions.STARTING_VERSION;
-import static io.delta.flink.source.DeltaSourceOptions.TIMESTAMP_AS_OF;
-import static io.delta.flink.source.DeltaSourceOptions.UPDATE_CHECK_INTERVAL;
-import static io.delta.flink.source.DeltaSourceOptions.VERSION_AS_OF;
+import static io.delta.flink.source.internal.DeltaSourceOptions.IGNORE_DELETES;
+import static io.delta.flink.source.internal.DeltaSourceOptions.PARQUET_BATCH_SIZE;
+import static io.delta.flink.source.internal.DeltaSourceOptions.STARTING_TIMESTAMP;
+import static io.delta.flink.source.internal.DeltaSourceOptions.STARTING_VERSION;
+import static io.delta.flink.source.internal.DeltaSourceOptions.TIMESTAMP_AS_OF;
+import static io.delta.flink.source.internal.DeltaSourceOptions.UPDATE_CHECK_INTERVAL;
+import static io.delta.flink.source.internal.DeltaSourceOptions.VERSION_AS_OF;
 
 public final class DeltaSourceBuilder {
