--- flink-connector/src/main/java/io/delta/flink/source/enumerator/SplitEnumeratorProvider.java
@@ -1,28 +1,0 @@
-package io.delta.flink.source.enumerator;
-
-import java.io.Serializable;
-
-import io.delta.flink.source.DeltaSourceOptions;
-import io.delta.flink.source.state.DeltaEnumeratorStateCheckpoint;
-import io.delta.flink.source.state.DeltaSourceSplit;
-import org.apache.flink.api.connector.source.Boundedness;
-import org.apache.flink.api.connector.source.SplitEnumerator;
-import org.apache.flink.api.connector.source.SplitEnumeratorContext;
-import org.apache.flink.core.fs.Path;
-import org.apache.hadoop.conf.Configuration;
-
-public interface SplitEnumeratorProvider extends Serializable {
-
-    SplitEnumerator<DeltaSourceSplit, DeltaEnumeratorStateCheckpoint<DeltaSourceSplit>>
-        createEnumerator(Path deltaTablePath, Configuration configuration,
-        SplitEnumeratorContext<DeltaSourceSplit> enumContext, DeltaSourceOptions sourceOptions);
-
-
-    SplitEnumerator<DeltaSourceSplit, DeltaEnumeratorStateCheckpoint<DeltaSourceSplit>>
-        createEnumerator(
-        DeltaEnumeratorStateCheckpoint<DeltaSourceSplit> checkpoint, Configuration configuration,
-        SplitEnumeratorContext<DeltaSourceSplit> enumContext, DeltaSourceOptions sourceOptions);
-
-    Boundedness getBoundedness();
-
-}
