--- flink-connector/src/main/java/io/delta/flink/source/file/AddFileEnumeratorContext.java
@@ -1,24 +1,0 @@
-package io.delta.flink.source.file;
-
-import java.util.List;
-
-import io.delta.standalone.actions.AddFile;
-
-public class AddFileEnumeratorContext {
-
-    private final String tablePath;
-    private final List<AddFile> addFiles;
-
-    public AddFileEnumeratorContext(String tablePath, List<AddFile> addFiles) {
-        this.tablePath = tablePath;
-        this.addFiles = addFiles;
-    }
-
-    public String getTablePath() {
-        return tablePath;
-    }
-
-    public List<AddFile> getAddFiles() {
-        return addFiles;
-    }
-}
