+++ flink-connector/src/main/java/io/delta/flink/source/internal/enumerator/SplitEnumeratorProvider.java
@@ -1,0 +1,28 @@
+package io.delta.flink.source.internal.enumerator;
+
+import java.io.Serializable;
+
+import io.delta.flink.source.internal.DeltaSourceOptions;
+import io.delta.flink.source.internal.state.DeltaEnumeratorStateCheckpoint;
+import io.delta.flink.source.internal.state.DeltaSourceSplit;
+import org.apache.flink.api.connector.source.Boundedness;
+import org.apache.flink.api.connector.source.SplitEnumerator;
+import org.apache.flink.api.connector.source.SplitEnumeratorContext;
+import org.apache.flink.core.fs.Path;
+import org.apache.hadoop.conf.Configuration;
+
+public interface SplitEnumeratorProvider extends Serializable {
+
+    SplitEnumerator<DeltaSourceSplit, DeltaEnumeratorStateCheckpoint<DeltaSourceSplit>>
+        createEnumerator(Path deltaTablePath, Configuration configuration,
+        SplitEnumeratorContext<DeltaSourceSplit> enumContext, DeltaSourceOptions sourceOptions);
+
+
+    SplitEnumerator<DeltaSourceSplit, DeltaEnumeratorStateCheckpoint<DeltaSourceSplit>>
+        createEnumerator(
+        DeltaEnumeratorStateCheckpoint<DeltaSourceSplit> checkpoint, Configuration configuration,
+        SplitEnumeratorContext<DeltaSourceSplit> enumContext, DeltaSourceOptions sourceOptions);
+
+    Boundedness getBoundedness();
+
+}
