+++ flink-connector/src/main/java/io/delta/flink/source/internal/exceptions/DeltaSourceException.java
@@ -1,0 +1,16 @@
+package io.delta.flink.source.internal.exceptions;
+
+public class DeltaSourceException extends RuntimeException {
+
+    public DeltaSourceException(Throwable cause) {
+        super(cause);
+    }
+
+    public DeltaSourceException(String message) {
+        super(message);
+    }
+
+    public DeltaSourceException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
