--- flink-connector/src/main/java/io/delta/flink/source/utils/SourceUtils.java
@@ -1,14 +1,0 @@
-package io.delta.flink.source.utils;
-
-import org.apache.flink.core.fs.Path;
-
-public final class SourceUtils {
-
-    private SourceUtils() {
-
-    }
-
-    public static String pathToString(Path path) {
-        return path.toUri().normalize().toString();
-    }
-}
