--- flink-connector/src/test/java/io/delta/flink/source/DeltaSourceBoundedExecutionITCase.java
@@ -1,218 +1,0 @@
-package io.delta.flink.source;
-
-import java.io.File;
-import java.io.IOException;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import io.delta.flink.DeltaTestUtils;
-import io.delta.flink.sink.utils.DeltaSinkTestUtils;
-import io.delta.flink.source.RecordCounterToFail.FailCheck;
-import org.apache.flink.core.fs.Path;
-import org.apache.flink.table.data.RowData;
-import org.apache.flink.table.types.logical.BigIntType;
-import org.apache.flink.table.types.logical.CharType;
-import org.apache.flink.table.types.logical.IntType;
-import org.apache.flink.table.types.logical.LogicalType;
-import org.apache.hadoop.conf.Configuration;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import static org.hamcrest.core.IsEqual.equalTo;
-import static org.junit.Assert.assertThat;
-
-public class DeltaSourceBoundedExecutionITCase extends DeltaSourceITBase {
-
-    private static final Set<String> EXPECTED_NAMES =
-        Stream.of("Kowalski", "Duda").collect(Collectors.toSet());
-
-    private static final Configuration HADOOP_CONF = DeltaTestUtils.getHadoopConf();
-
-    private static final int LARGE_TABLE_COUNT = 1100;
-
-    private String nonPartitionedTablePath;
-
-    private String nonPartitionedLargeTablePath;
-
-    private String partitionedTablePath;
-
-    @Before
-    public void setup() {
-        try {
-            nonPartitionedTablePath = TMP_FOLDER.newFolder().getAbsolutePath();
-            partitionedTablePath = TMP_FOLDER.newFolder().getAbsolutePath();
-            nonPartitionedLargeTablePath = TMP_FOLDER.newFolder().getAbsolutePath();
-
-            DeltaSinkTestUtils.initTestForNonPartitionedTable(nonPartitionedTablePath);
-            DeltaSinkTestUtils.initTestForPartitionedTable(partitionedTablePath);
-            DeltaSinkTestUtils.initTestForNonPartitionedLargeTable(
-                nonPartitionedLargeTablePath);
-        } catch (IOException e) {
-            throw new RuntimeException("Weren't able to setup the test dependencies", e);
-        }
-    }
-
-    @After
-    public void after() {
-        miniClusterResource.getClusterClient().close();
-    }
-
-    @Test
-    public void testWithoutPartitions() throws Exception {
-
-        // GIVEN
-        DeltaSource<RowData> deltaSource = DeltaSourceBuilder.builder()
-            .tablePath(Path.fromLocalFile(new File(nonPartitionedTablePath)))
-            .columnNames(new String[]{"name", "surname", "age"})
-            .columnTypes(new LogicalType[]{new CharType(), new CharType(), new IntType()})
-            .hadoopConfiguration(DeltaTestUtils.getHadoopConf())
-            .buildForRowData();
-
-        // WHEN
-        List<RowData> resultData = testBoundDeltaSource(deltaSource);
-
-        Set<String> actualNames =
-            resultData.stream().map(row -> row.getString(1).toString()).collect(Collectors.toSet());
-
-        // THEN
-        assertThat("Source read different number of rows that Delta Table have.", resultData.size(),
-            equalTo(2));
-        assertThat("Source Produced Different Rows that were in Delta Table", actualNames,
-            equalTo(EXPECTED_NAMES));
-
-        //System.out.println(resultData);
-    }
-
-    @Test
-    public void testWithOnePartitions() throws Exception {
-
-        // GIVEN
-        DeltaSource<RowData> deltaSource = DeltaSourceBuilder.builder()
-            .tablePath(Path.fromLocalFile(new File(partitionedTablePath)))
-            .columnNames(new String[]{"name", "surname", "age", "col2"})
-            .columnTypes(
-                new LogicalType[]{new CharType(), new CharType(), new IntType(), new CharType()})
-            .hadoopConfiguration(DeltaTestUtils.getHadoopConf())
-            .partitions(Arrays.asList("col1", "col2"))
-            .buildForRowData();
-
-        // WHEN
-        List<RowData> resultData = testBoundDeltaSource(deltaSource);
-
-        Set<String> actualNames =
-            resultData.stream().map(row -> row.getString(1).toString()).collect(Collectors.toSet());
-
-        // THEN
-        assertThat("Source read different number of rows that Delta Table have.", resultData.size(),
-            equalTo(2));
-        assertThat("Source Produced Different Rows that were in Delta Table", actualNames,
-            equalTo(EXPECTED_NAMES));
-
-        resultData.forEach(rowData -> {
-            assertPartitionValue(rowData, 3, "val2");
-        });
-
-        //System.out.println(resultData);
-    }
-
-    @Test
-    public void testWithBothPartitions() throws Exception {
-
-        // GIVEN
-        DeltaSource<RowData> deltaSource = DeltaSourceBuilder.builder()
-            .tablePath(Path.fromLocalFile(new File(partitionedTablePath)))
-            .columnNames(new String[]{"name", "surname", "age", "col1", "col2"})
-            .columnTypes(
-                new LogicalType[]{new CharType(), new CharType(), new IntType(), new CharType(),
-                    new CharType()})
-            .hadoopConfiguration(DeltaTestUtils.getHadoopConf())
-            .partitions(Arrays.asList("col1", "col2"))
-            .buildForRowData();
-
-        // WHEN
-        List<RowData> resultData = testBoundDeltaSource(deltaSource);
-
-        Set<String> actualNames =
-            resultData.stream().map(row -> row.getString(1).toString()).collect(Collectors.toSet());
-
-        // THEN
-        assertThat("Source read different number of rows that Delta Table have.", resultData.size(),
-            equalTo(2));
-        assertThat("Source Produced Different Rows that were in Delta Table", actualNames,
-            equalTo(EXPECTED_NAMES));
-
-        resultData.forEach(rowData -> {
-            assertPartitionValue(rowData, 3, "val1");
-            assertPartitionValue(rowData, 4, "val2");
-        });
-
-        //System.out.println(resultData);
-    }
-
-    @Test
-    // NOTE that this test can take some time to finish since we are restarting JM here.
-    // It can be around 30 seconds or so.
-    // Test if SplitEnumerator::addSplitsBack works well,
-    // meaning if splits were added back to the Enumerator's state and reassigned to new TM.
-    public void testWithTaskManagerFailover() throws Exception {
-
-        // GIVEN
-        DeltaSource<RowData> deltaSource = DeltaSourceBuilder.builder()
-            .tablePath(Path.fromLocalFile(new File(nonPartitionedLargeTablePath)))
-            .columnNames(new String[]{"col1", "col2", "col3"})
-            .columnTypes(new LogicalType[]{new BigIntType(), new BigIntType(), new CharType()})
-            .hadoopConfiguration(DeltaTestUtils.getHadoopConf())
-            .buildForRowData();
-
-        // WHEN
-        List<RowData> resultData = testBoundDeltaSource(FailoverType.TM, deltaSource,
-            (FailCheck) readRows -> readRows == LARGE_TABLE_COUNT / 2);
-
-        Set<Long> actualValues =
-            resultData.stream().map(row -> row.getLong(0)).collect(Collectors.toSet());
-
-        // THEN
-        assertThat("Source read different number of rows that Delta Table have.", resultData.size(),
-            equalTo(LARGE_TABLE_COUNT));
-        assertThat("Source Must Have produced some duplicates.", actualValues.size(),
-            equalTo(LARGE_TABLE_COUNT));
-
-        //System.out.println(CountSink.count.get());
-        //System.out.println(resultData);
-    }
-
-    @Test
-    public void testWithJobManagerFailover() throws Exception {
-        // GIVEN
-        DeltaSource<RowData> deltaSource = DeltaSourceBuilder.builder()
-            .tablePath(Path.fromLocalFile(new File(nonPartitionedLargeTablePath)))
-            .columnNames(new String[]{"col1", "col2", "col3"})
-            .columnTypes(new LogicalType[]{new BigIntType(), new BigIntType(), new CharType()})
-            .hadoopConfiguration(DeltaTestUtils.getHadoopConf())
-            .buildForRowData();
-
-        // WHEN
-        List<RowData> resultData = testBoundDeltaSource(FailoverType.JM, deltaSource,
-            (FailCheck) readRows -> readRows == LARGE_TABLE_COUNT / 2);
-
-        Set<Long> actualValues =
-            resultData.stream().map(row -> row.getLong(0)).collect(Collectors.toSet());
-
-        // THEN
-        assertThat("Source read different number of rows that Delta Table have.", resultData.size(),
-            equalTo(LARGE_TABLE_COUNT));
-        assertThat("Source Must Have produced some duplicates.", actualValues.size(),
-            equalTo(LARGE_TABLE_COUNT));
-
-        //System.out.println(resultData);
-    }
-
-
-    private void assertPartitionValue(RowData rowData, int pos, String val1) {
-        assertThat("Partition column has a wrong value.", rowData.getString(pos).toString(),
-            equalTo(val1));
-    }
-}
